utils/backend_connector.py
"""
Backend connector for Databricks integration
Handles communication with existing Databricks provider verification backend
"""

import requests
import os
from datetime import datetime


# Databricks configuration (set via environment variables)
DATABRICKS_HOST = os.environ.get('DATABRICKS_HOST', 'https://your-workspace.cloud.databricks.com')
DATABRICKS_TOKEN = os.environ.get('DATABRICKS_TOKEN', '')
BATCH_JOB_ID = os.environ.get('BATCH_JOB_ID', '')


def process_batch(csv_path, batch_id):
    """
    Trigger batch processing on Databricks backend.
    
    This function interfaces with your existing Databricks backend that performs:
    - AI-powered provider verification
    - Confidence scoring
    - Data source consultation
    - Discrepancy detection
    
    Args:
        csv_path: Path to uploaded CSV/Excel file
        batch_id: Unique identifier for this batch
        
    Returns:
        str: Batch ID for tracking processing status
        
    Implementation Options:
    ----------------------
    Option 1: Databricks Jobs API
        - Trigger a notebook or job run
        - Pass file path as parameter
        
    Option 2: REST API Endpoint
        - POST to custom API endpoint
        - Backend processes and stores results
        
    Option 3: Delta Table Write + Monitoring
        - Write input to Delta table
        - Monitor output table for results
    """
    
    # TODO: Implement based on your backend architecture
    # Choose one of the options below:
    
    # --- OPTION 1: Databricks Jobs API ---
    # Uncomment and configure when ready:
    """
    response = requests.post(
        f"{DATABRICKS_HOST}/api/2.1/jobs/run-now",
        headers={"Authorization": f"Bearer {DATABRICKS_TOKEN}"},
        json={
            "job_id": BATCH_JOB_ID,
            "notebook_params": {
                "csv_path": csv_path,
                "batch_id": batch_id
            }
        }
    )
    
    if response.status_code == 200:
        run_id = response.json()['run_id']
        print(f"Batch {batch_id} submitted. Run ID: {run_id}")
        return batch_id
    else:
        raise Exception(f"Failed to submit batch: {response.text}")
    """
    
    # --- OPTION 2: Custom REST API ---
    # Uncomment when your backend exposes an API:
    """
    api_url = f"{DATABRICKS_HOST}/api/custom/verify-providers"
    response = requests.post(
        api_url,
        headers={"Authorization": f"Bearer {DATABRICKS_TOKEN}"},
        json={
            "csv_path": csv_path,
            "batch_id": batch_id
        }
    )
    
    if response.status_code == 200:
        return batch_id
    else:
        raise Exception(f"Backend API error: {response.text}")
    """
    
    # --- OPTION 3: Delta Table Integration ---
    # Uncomment when using Spark/Delta:
    """
    from pyspark.sql import SparkSession
    import pandas as pd
    
    spark = SparkSession.builder.getOrCreate()
    
    # Read CSV and write to input Delta table
    df = pd.read_csv(csv_path)
    spark_df = spark.createDataFrame(df)
    spark_df.write \
        .format("delta") \
        .mode("append") \
        .option("mergeSchema", "true") \
        .saveAsTable("provider_verification_input")
    
    print(f"Batch {batch_id} written to Delta table")
    return batch_id
    """
    
    # For now, just return the batch_id (mock mode)
    print(f"[MOCK MODE] Would trigger backend processing for batch {batch_id}")
    return batch_id


def get_batch_results(batch_id):
    """
    Retrieve verification results for a batch from Databricks backend.
    
    Args:
        batch_id: Unique identifier for the batch
        
    Returns:
        list: List of dictionaries containing verification results
        
    Expected result format:
    [
        {
            'provider_name': 'Dr. John Smith',
            'npi': '1234567890',
            'original_address': '123 Main St',
            'verified_address': '123 Main St',
            'specialty': 'Cardiology',
            'confidence_score': 95,
            'status': 'verified',
            'sources': ['NPPES', 'Medical Board'],
            'address_changed': False,
            ...
        },
        ...
    ]
    """
    
    # TODO: Implement based on your backend architecture
    # Choose one of the options below:
    
    # --- OPTION 1: REST API ---
    """
    api_url = f"{DATABRICKS_HOST}/api/custom/batch/{batch_id}/results"
    response = requests.get(
        api_url,
        headers={"Authorization": f"Bearer {DATABRICKS_TOKEN}"}
    )
    
    if response.status_code == 200:
        return response.json()['results']
    else:
        raise Exception(f"Failed to fetch results: {response.text}")
    """
    
    # --- OPTION 2: Delta Table Query ---
    """
    from pyspark.sql import SparkSession
    
    spark = SparkSession.builder.getOrCreate()
    results_df = spark.table("provider_verification_results") \
                     .filter(f"batch_id = '{batch_id}'") \
                     .toPandas()
    
    return results_df.to_dict('records')
    """
    
    # --- OPTION 3: Databricks SQL ---
    """
    from databricks import sql
    
    connection = sql.connect(
        server_hostname=DATABRICKS_HOST,
        http_path=os.environ.get('DATABRICKS_HTTP_PATH'),
        access_token=DATABRICKS_TOKEN
    )
    
    cursor = connection.cursor()
    cursor.execute(f'''
        SELECT * FROM provider_verification_results
        WHERE batch_id = '{batch_id}'
    ''')
    
    results = cursor.fetchall()
    columns = [desc[0] for desc in cursor.description]
    
    return [dict(zip(columns, row)) for row in results]
    """
    
    # For now, raise an error to force mock mode
    raise NotImplementedError("Backend integration not yet configured. Using mock data mode.")


def check_batch_status(batch_id):
    """
    Check if batch processing is complete.
    
    Args:
        batch_id: Unique identifier for the batch
        
    Returns:
        dict: Status information
            {
                'status': 'pending' | 'processing' | 'completed' | 'failed',
                'progress': 0-100,
                'message': 'Status message'
            }
    """
    
    # TODO: Implement based on your backend
    # This is useful for showing progress while processing large batches
    
    """
    # Example: Check Databricks job run status
    response = requests.get(
        f"{DATABRICKS_HOST}/api/2.1/jobs/runs/get",
        headers={"Authorization": f"Bearer {DATABRICKS_TOKEN}"},
        params={"run_id": batch_id}
    )
    
    if response.status_code == 200:
        run_info = response.json()
        state = run_info['state']['life_cycle_state']
        
        status_map = {
            'PENDING': 'pending',
            'RUNNING': 'processing',
            'TERMINATING': 'processing',
            'TERMINATED': 'completed',
            'INTERNAL_ERROR': 'failed'
        }
        
        return {
            'status': status_map.get(state, 'unknown'),
            'progress': 50 if state == 'RUNNING' else 100,
            'message': f"Job {state.lower()}"
        }
    """
    
    return {
        'status': 'completed',
        'progress': 100,
        'message': 'Processing complete (mock mode)'
    }
